app/requirements.txt
feedparser
PyYAML
SQLAlchemy
psycopg2-binary
requests
openai
APScheduler
Flask
langchain
